pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Railway template for fast-api with mysql"
authors = ["rschwemmer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic= "1.11.3"
annotated-types= "0.5.0"
anyio= "3.7.1"
bcrypt= "4.0.1"
click= "8.1.7"
colorama= "0.4.6"
ecdsa= "0.18.0"
fastapi= "0.103.0"
FastAPI-SQLAlchemy= "0.2.1"
greenlet= "3.0.3"
h11= "0.14.0"
idna= "3.4"
Mako= "1.2.4"
MarkupSafe= "2.1.3"
passlib= "1.7.4"
pydantic= "2.5.2"
pydantic-extra-types= "2.4.0"
pydantic-settings= "2.0.3"
pydantic_core= "2.14.5"
PyMySQL= "1.1.0"
python-dotenv= "1.0.0"
python-jose= "3.3.0"
rsa= "4.9"
six= "1.16.0"
sniffio= "1.3.0"
SQLAlchemy= "2.0.20"
starlette= "0.27.0"
typing_extensions= "4.7.1"
uvicorn= "0.23.2"


[tool.poetry.scripts]
dev = "main:dev"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
